filter out unplayable/unavailable tracks (#207)
Resolves #158.
Resolves #211.

The purpose of this PR is to filter out unavailable songs. This functionality is achieved by utilizing Spotify's song relinking feature, specifically by sending the `market` query parameter and using `from_token` to retrieve the user's country code from their token.

A track is filtered out if its `is_playable` is `false`.

This did require a non significant refactor with the functions that call ``internal_call`` and the various internal pagination functions. But by making a "generic" payload field allows for more flexibility in the future if Spotify adds more query params.