ci(deps): Update github actions
Migrate from actions-rs/* to dtolnay/*, update other actions to latest
versions.

The actions-rs was archived not long ago, and while the actual result
will not change, this prevents github action warnings.

Also reorders the actions to have an early failure with cargo fmt, and
added the beta toolchain to the CI workflow

In CD use checkout@v3 to avoid missing glibc issues