recommend official Arch package instead of AUR
An official Arch package has been available since 2025-02-05. Since this
is the simplest way to install packages on Arch, it makes sense for most
users to prefer this over the AUR packages. The AUR package that is
currently recommended in the documentation is also no longer available
because of the adoption in the official repositories.