add display of genres in playback UI via `{genres}` (#795)
As suggested in #145 this pull request adds the functionality to display genres in the playback UI.

The number of genres is adjustable via the `genre_num` config option.
`{genres}` can be used in the `playback_format` config to place the genres where wished.

The genres are formatted as follows (`genre_num = 2`)
- 1 genre: "genre1"
- 2 genres: "genre1, genre2"
- \>= 3 genres: "genre1, genre2, ..."

Here is an example with the default `genre_num = 2`
<img width="1118" height="238" alt="2025-08-05-213208" src="https://github.com/user-attachments/assets/3f2f429a-1336-4607-bad7-0e1c40891dee" />

The genres are retrieved for the first/main artist of a track and cached in association to this artist.

closes #145 

---------

Co-authored-by: Thang Pham <phamducthang1234@gmail.com>