feat(nix): add nix flake
This way the package can be build and run using nix commands and the
flake can be used to get a more up-to-date version of spotify-player
than nixpkgs has. This can be particularly useful to nix users if there
is a fix on master which hasn't made its way downstream.